refactored so that defaults.ts to contains LAYER_DEFAULTS
