+++ giraffe/src/constants/index.ts
@@ -9,8 +9,7 @@ import {
   curveNatural,
 } from 'd3-shape'
 
-import {Config, LayerConfig, SymbolType} from '../types'
-import {NINETEEN_EIGHTY_FOUR as DEFAULT_COLOR_SCHEME} from './colorSchemes'
+import {Config, SymbolType} from '../types'
 
 // TODO: Make configurable
 export const TICK_PADDING_RIGHT = 8
@@ -53,52 +52,6 @@ export const CONFIG_DEFAULTS: Partial<Config> = {
   legendBorder: '2px solid #202028',
   legendCrosshairColor: '#31313d',
   legendColorizeRows: true,
-}
-
-export const LAYER_DEFAULTS: {[layerType: string]: Partial<LayerConfig>} = {
-  line: {
-    lineWidth: 1,
-    hoverDimension: 'auto',
-    fill: [],
-    colors: DEFAULT_COLOR_SCHEME,
-    interpolation: 'linear',
-    maxTooltipRows: 24,
-    shadeBelow: false,
-    shadeBelowOpacity: 0.1,
-  },
-  heatmap: {
-    colors: DEFAULT_COLOR_SCHEME,
-    binSize: 10,
-    strokeWidth: 0,
-    strokePadding: 0,
-    strokeOpacity: 0,
-    fillOpacity: 1,
-  },
-  scatter: {
-    colors: DEFAULT_COLOR_SCHEME,
-    fill: [],
-    symbol: [],
-  },
-  histogram: {
-    fill: [],
-    colors: DEFAULT_COLOR_SCHEME,
-    position: 'stacked',
-    binCount: null,
-    strokeWidth: 1,
-    strokePadding: 0.75,
-    strokeOpacity: 1,
-    fillOpacity: 0.75,
-  },
-  mosaic: {
-    fill: [],
-    colors: DEFAULT_COLOR_SCHEME,
-    position: 'stacked',
-    binCount: null,
-    strokeWidth: 1,
-    strokePadding: 0.75,
-    strokeOpacity: 1,
-    fillOpacity: 0.75,
-  },
 }
 
 export const ALL_SYMBOL_TYPES: SymbolType[] = [
